Add an interface for the JiraQueryHandler
