+++ src/test/java/de/uhd/ifi/se/decision/management/jira/filtering/TestJiraQueryHandler.java
@@ -12,6 +12,7 @@ import com.atlassian.jira.component.ComponentAccessor;
 import com.atlassian.jira.user.ApplicationUser;
 
 import de.uhd.ifi.se.decision.management.jira.TestSetUpWithIssues;
+import de.uhd.ifi.se.decision.management.jira.filtering.impl.JiraQueryHandlerImpl;
 
 public class TestJiraQueryHandler extends TestSetUpWithIssues {
 
@@ -26,27 +27,28 @@ public class TestJiraQueryHandler extends TestSetUpWithIssues {
 
 	@Test
 	public void testConstructorQueryNull() {
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", null);
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST", null);
 		assertEquals(JiraQueryType.OTHER, jiraQueryHandler.getQueryType());
 		assertEquals("type = null", jiraQueryHandler.getQuery());
 	}
 
 	@Test
 	public void testConstructorQueryEmpty() {
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", "");
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST", "");
 		assertEquals(JiraQueryType.OTHER, jiraQueryHandler.getQueryType());
 		assertEquals("type = null", jiraQueryHandler.getQuery());
 	}
 
 	@Test
 	public void testGetJiraIssuesFromEmptyQuery() {
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", "");
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST", "");
 		assertEquals(0, jiraQueryHandler.getJiraIssuesFromQuery().size());
 	}
 
 	@Test
 	public void testGetNamesOfJiraIssueTypesInQueryOneType() {
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", "?jql=project=TEST AND issuetype = Issue AND resolution = Unresolved");
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST",
+				"?jql=project=TEST AND issuetype = Issue AND resolution = Unresolved");
 		assertEquals(JiraQueryType.JQL, jiraQueryHandler.getQueryType());
 		assertEquals("project=TEST AND issuetype = Issue AND resolution = Unresolved", jiraQueryHandler.getQuery());
 		List<String> types = jiraQueryHandler.getNamesOfJiraIssueTypesInQuery();
@@ -56,16 +58,17 @@ public class TestJiraQueryHandler extends TestSetUpWithIssues {
 
 	@Test
 	public void testGetNamesOfJiraIssueTypesInQueryZeroTypes() {
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", "?jql=project=TEST");
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST", "?jql=project=TEST");
 		List<String> types = jiraQueryHandler.getNamesOfJiraIssueTypesInQuery();
 		assertEquals(0, types.size());
 	}
 
 	@Test
 	public void testGetNamesOfJiraIssueTypesInQueryThreeTypes() {
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST",
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST",
 				"abc§jql=project=TEST AND issuetype in (Decision, Issue, Alternative) AND resolution = Unresolved");
-		assertEquals("project=TEST AND issuetype in (Decision, Issue, Alternative) AND resolution = Unresolved", jiraQueryHandler.getQuery());
+		assertEquals("project=TEST AND issuetype in (Decision, Issue, Alternative) AND resolution = Unresolved",
+				jiraQueryHandler.getQuery());
 		List<String> types = jiraQueryHandler.getNamesOfJiraIssueTypesInQuery();
 		assertEquals("Decision", types.get(0));
 		assertEquals("Issue", types.get(1));
@@ -75,7 +78,7 @@ public class TestJiraQueryHandler extends TestSetUpWithIssues {
 
 	@Test
 	public void testGetDatesNoDatesInQuery() {
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", "");
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST", "");
 		assertEquals(-1, jiraQueryHandler.getCreatedEarliest());
 		assertEquals(-1, jiraQueryHandler.getCreatedLatest());
 	}
@@ -83,30 +86,28 @@ public class TestJiraQueryHandler extends TestSetUpWithIssues {
 	@Test
 	public void testGetDatesInQuery() {
 		assertTrue("1970-01-01".matches("\\d\\d\\d\\d-\\d\\d-\\d\\d(.)*"));
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", "?jql=resolution = Unresolved AND created >= 1970-01-02 AND created <= 1970-01-03");
-		assertEquals("resolution = Unresolved AND created >= 1970-01-02 AND created <= 1970-01-03", jiraQueryHandler.getQuery());
-		assertEquals("created >= 1970-01-02 AND created <= 1970-01-03", jiraQueryHandler.getQuerySubstringWithTimeInformation());
-		assertEquals("1970-01-02 AND created <= 1970-01-03", jiraQueryHandler.getQuerySubstringWithTimeInformation().split(">=")[1].trim());
-		
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST",
+				"?jql=resolution = Unresolved AND created >= 1970-01-02 AND created <= 1970-01-03");
+		assertEquals("resolution = Unresolved AND created >= 1970-01-02 AND created <= 1970-01-03",
+				jiraQueryHandler.getQuery());
 		assertTrue(jiraQueryHandler.getCreatedEarliest() > 0);
 		assertTrue(jiraQueryHandler.getCreatedLatest() > 0);
 	}
 
 	@Test
 	public void testGetDatesInQueryFromTimeFactor() {
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", "?jql=created >= -1m AND created <= -1w");
-		assertEquals("created >= -1m AND created <= -1w", jiraQueryHandler.getQuerySubstringWithTimeInformation());
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST", "?jql=created >= -1m AND created <= -1w");
 		assertTrue(jiraQueryHandler.getCreatedEarliest() > 100000000);
 		assertTrue(jiraQueryHandler.getCreatedLatest() > 100000000);
 
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", "?jql=created >= -1d AND created <= -1h");
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST", "?jql=created >= -1d AND created <= -1h");
 		assertTrue(jiraQueryHandler.getCreatedEarliest() > 100000000);
 		assertTrue(jiraQueryHandler.getCreatedLatest() > 100000000);
 	}
 
 	@Test
 	public void testGetQueryObject() {
-		jiraQueryHandler = new JiraQueryHandler(user, "TEST", "?jql=project");
+		jiraQueryHandler = new JiraQueryHandlerImpl(user, "TEST", "?jql=project");
 		assertEquals("project", jiraQueryHandler.getQueryObject().getQueryString());
 	}
